add patch for forward ref including when it's wrapped with memo
